+++ src/main/java/de/wevsvirushackathon/coronareport/healthdepartment/HealthDepartmentRepository.java
@@ -6,6 +6,4 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface HealthDepartmentRepository
         extends CrudRepository<HealthDepartment, String> {
-
-
 }
